Added support for ISUPPORT
This commit adds support for using the RPL_005 (ISUPPORT) reported
Feature Details of a server. The main thing is to quit using hardcoded
Prefix Modes in the code by parsing out ISUPPORT's PREFIX Parameter as
well as pulling out the correct NickLength.

This uses the defaults according to ircdocs.horse and replaces them with
what the server sends to us.

This adds in a few members to the Connection struct:
  Features
  KnownFeatures
  featuresMutex

The general idea is that only members of the package will modify these
fields and will be sure to take out a lock against the one Mutex when
doing so.

Features and KnownFeatures are for client consumption.