Change how synchronization is done.
Prior we were worried about parallel reads/writes to maps within structs
but upon further checking we've realized not only need we protect
against that we also need handlers to run synchronously (one after the
other) and this change should provide for that. We've swapped to using a
RWMutex and removed all granular locking.